refactor: move login service outside services/services
